requirements.txt
fastapi
uvicorn
python-dotenv
openai-whisper
ffmpeg-python
pyttsx3
python-multipart